backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "boto3>=1.42.59",
    "fastapi>=0.135.1",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.3.0",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]
